Listo en Agregar/Modificar/Eliminar Cirugia y Paciente

+++ BD_Grupo3_VS/BD_Grupo3_VS/VerPaciente.cs
@@ -24,6 +24,7 @@ namespace BD_Grupo3_VS
         bool cambiosFechaNac = false;
         bool cambiosValoracion = false;
         bool cambiosComentarios = false;
+        bool cambiosCirugia = false;
         string nombre;
         string apellido1;
         string apellido2;
@@ -68,21 +69,28 @@ namespace BD_Grupo3_VS
             dato.Read();
             DTP_FechaNac.Value = dato.GetDateTime(0);
 
+            llenarTabla(DGV_Cirugias);
+
         }
 
 
         private void Inicio()
         {
 
-        }    
+        }
 
         private void BTN_Eliminar_Click(object sender, EventArgs e)
         {
-            paciente.eliminarPaciente(cedula);
-            MessageBox.Show("¡El paciente ha sido agregado exitosamente!", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.None);
-            MenuPrincipal menu = new MenuPrincipal();
-            menu.Show();
-            this.Hide();
+            MessageBoxButtons botones = MessageBoxButtons.YesNo;
+            DialogResult resultado = MessageBox.Show("¿Seguro que desea elimar al paciente?", "Eliminar Paciente", botones);
+            if (resultado == System.Windows.Forms.DialogResult.Yes)
+            {
+                paciente.eliminarPaciente(cedula);
+                MessageBox.Show("¡El paciente ha sido eliminado exitosamente!", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.None);
+                MenuPrincipal menu = new MenuPrincipal();
+                menu.Show();
+                this.Hide();
+            }
         }
 
         private void BTN_Modificar_Click(object sender, EventArgs e)
@@ -110,7 +118,7 @@ namespace BD_Grupo3_VS
             if (cambiosApellido1)
             {
                 result = paciente.modificarApellido1(TXT_Cedula.Text, TXT_Apellido1.Text);
-                if(result != 0)
+                if (result != 0)
                 {
                     MessageBox.Show("Ha ocurrido un error.", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     cambioshechos = false;
@@ -153,6 +161,15 @@ namespace BD_Grupo3_VS
                     cambioshechos = false;
                 }
             }
+            if (cambiosSexo)
+            {
+                result = paciente.modificarSexo(TXT_Cedula.Text, TXT_Sexo.Text);
+                if (result != 0)
+                {
+                    MessageBox.Show("Ha ocurrido un error.", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    cambioshechos = false;
+                }
+            }
             if (cambiosValoracion)
             {
                 result = paciente.modificarValoracion(TXT_Cedula.Text, TXT_Valoracion.Text);
@@ -180,10 +197,9 @@ namespace BD_Grupo3_VS
                 VerPaciente vp = new VerPaciente(TXT_Cedula.Text, TXT_Nombre.Text, TXT_Apellido1.Text, TXT_Apellido2.Text);
                 this.Dispose();
                 vp.Show();
-                
             }
         }
-    
+
 
         private void TXT_Nombre_TextChanged(object sender, EventArgs e)
         {
@@ -195,7 +211,7 @@ namespace BD_Grupo3_VS
         }
 
         private void TXT_Apellido1_TextChanged(object sender, EventArgs e)
-        {           
+        {
             if (cambiosApellido1)
             {
                 TXT_Apellido1.BackColor = System.Drawing.Color.LightBlue;
@@ -204,7 +220,7 @@ namespace BD_Grupo3_VS
         }
 
         private void TXT_Apellido2_TextChanged(object sender, EventArgs e)
-        {           
+        {
             if (cambiosApellido2)
             {
                 TXT_Apellido2.BackColor = System.Drawing.Color.LightBlue;
@@ -274,7 +290,7 @@ namespace BD_Grupo3_VS
             {
                 TXT_Comentarios.BackColor = System.Drawing.Color.LightBlue;
             }
-                cambiosComentarios = true;           
+            cambiosComentarios = true;
         }
 
         /*             A partir de aqui empiezan los metodos para la cinta del menu  */
@@ -317,5 +333,109 @@ namespace BD_Grupo3_VS
             }
         }
         /*             Hasta aqui las instrucciones de la cinta del menu  */
+
+        private void BTN_AgregarCirugia_Click(object sender, EventArgs e)
+        {
+            int resultado = paciente.agregarCirugia(TXT_Cedula.Text, TXT_NuevaCirugia.Text);
+
+            //resultado es 0 cuando se pudo agregar un antecedente al paciente con éxito
+            if (resultado == 0)
+            {
+                MessageBox.Show("¡La cirugía ha sido agregado exitosamente!", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.None);
+                DGV_Cirugias.Refresh();
+            }
+            else
+            {
+                if (resultado == 2627)
+                {
+                    MessageBox.Show("Ya existe esta cirugía asociado al paciente", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+                else
+                {
+                    MessageBox.Show(resultado + ": Algo ha fallado.", "Oops", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+            }
+        }
+
+        private void llenarTabla(DataGridView dataGridView)
+        {
+
+            DataTable tabla = paciente.obtenerCirugias(TXT_Cedula.Text);
+            BindingSource bindingSource = new BindingSource();
+            bindingSource.DataSource = tabla;
+            dataGridView.AutoResizeColumns(DataGridViewAutoSizeColumnsMode.AllCellsExceptHeader);
+            dataGridView.DataSource = bindingSource;
+            for (int i = 0; i < DGV_Cirugias.ColumnCount; i++)
+            {
+                dataGridView.Columns[i].Width = 100;
+            }
+        }
+
+        private void TXT_CirugiaSeleccionada_TextChanged(object sender, EventArgs e)
+        {
+            if (cambiosCirugia)
+            {
+                TXT_CirugiaSeleccionada.BackColor = System.Drawing.Color.LightBlue;
+            }
+            cambiosCirugia = true;
+        }
+
+        private void BTN_ModificarCirugia_Click(object sender, EventArgs e)
+        {
+            int result;
+            DataGridViewRow row = DGV_Cirugias.CurrentRow;
+            string cirugiaVieja = row.Cells[0].Value.ToString();
+            if (cambiosCirugia)
+            {
+                result = paciente.modificarCirugia(cedula, cirugiaVieja, TXT_CirugiaSeleccionada.Text);
+                if (result == 0)
+                {
+                    MessageBox.Show("¡La cirugía ha sido modificada exitosamente!", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.None);
+                    //TXT_CirugiaSeleccionada.Clear();
+                    VerPaciente vp = new VerPaciente(TXT_Cedula.Text, TXT_Nombre.Text, TXT_Apellido1.Text, TXT_Apellido2.Text);
+                    this.Dispose();
+                    vp.Show();
+                }
+                else
+                {
+                    if (result == 2627)
+                    {
+                        MessageBox.Show("Ya existe esta cirugía asociada al paciente. Por favor ingrese otra.", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        MessageBox.Show("Ha ocurrido un error.", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+
+                }
+            }
+        }
+
+        private void BTN_EliminarCirugia_Click(object sender, EventArgs e)
+        {
+            DataGridViewRow row = DGV_Cirugias.CurrentRow;
+            string cirugia = row.Cells[0].Value.ToString();
+            int result = paciente.eliminarCirugia(cedula, cirugia);
+            if (result == 0)
+            {
+                MessageBox.Show("¡La cirugía ha sido eliminada exitosamente!", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.None);
+                VerPaciente vp = new VerPaciente(TXT_Cedula.Text, TXT_Nombre.Text, TXT_Apellido1.Text, TXT_Apellido2.Text);
+                this.Dispose();
+                vp.Show();
+            }
+            else
+            {
+                MessageBox.Show("Ha ocurrido un error.", "Resultados", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+        }
+
+
+        private void DGV_Cirugias_CellContentDoubleClick(object sender, DataGridViewCellEventArgs e)
+        {
+            DataGridViewRow row = DGV_Cirugias.CurrentRow;
+            string cirugia = row.Cells[0].Value.ToString();
+            TXT_CirugiaSeleccionada.Text = cirugia;
+        }
     }
 }
+
